tmp/api/requirements.txt
autopep8==1.4
backports-abc==0.5
boto3==1.9.71
botocore==1.12.71
docutils==0.14
expiringdict==1.1.4
functools32==3.2.3.post2
futures==3.2.0
jmespath==0.9.3
jsonschema==2.6.0
psycopg2==2.7.5
pycodestyle==2.4.0
python-dateutil==2.7.3
PyYAML==3.13
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.2.11
tornado==5.1
urllib3==1.23
PyJWT==1.6.4
cryptography==2.4.2